Add files via upload
This function serves to migrate a Climada portfolio and model to a ktools model. From Climada we need the 'entity' structure and the 'hazard' structure which are the relevant input parameters of this code. This code is written as an example for a tropicaly cyclone wind. It is meant as a starting point for migrating any model to ktools/Oasis. But there must be code changes depending on the hazard, binning, portfolio etc.

The output of this code is a set of CSV files which are as specified by the Oasis team in the ktools documentation (GitHub). These CSV files are written into a folder in the user's home directory, into a sub-folder named below in the code ('ktools_tcusa' by default). 

From there they can be converted to binary files (.bin) with the standard ktools functions and the loss simulation can be done. We give an example call at the end of this code, but these are as documented in the ktools GitHub.

It is important to note that we do not expect this code to run for all po