take into account ROW naming in unaggregated case
We either have multiple RoW's (Rest of World Europe, ...) or one aggregated RoW in our mriot table. In the latter case the naming is still 'ROW' and not 'RoW'. We have to take into account this naming inconsistency so calculations are done properly. In future, we might find a better solution.