Ghosh methodology implemented
From now on it is possible to choose between the Ghosh methodology (=default) that is an extension to the standard IO model. It estimated indirect risk with the idea of a supply-driven economy. The user still has the option to apply the standard IO model (demand-driven) or EEIOA ('environmental accounting').