IO_YDS as output/input
IO_YDS is a climada Input-Output Year damage set struct that contains information on the direct (IO_YDS.direct) and indirect (IO_YDS.indirect) damage calculated. The IO_YDS is based on the sectorial representation of assets and hazard provided that are combined to obtain a YDS for each sector and country-combination. The (direct) year damage set is then stored in the IO_YDS.direct and used to calculate the indirect year damage set IO_YDS.indirect using Input-Output methodology.